Historic day. Weights perfectly fully transfered over! Use pspnet.py to run
